+++ shuup/front/utils/dashboard.py
@@ -29,6 +29,11 @@ class DashboardItem(object):
     title = _("Dashboard")
 
     """
+    Sequence number of dashboard item
+    """
+    ordering = 0
+
+    """
     The size of dashboard item block
     """
     size = DashboardItemSize.LARGE